Code tidy up based on sonarqube output
