+++ src/AcceptanceTests/TestContext.cs
@@ -1,14 +1,12 @@
-﻿using SFA.DAS.Funding.ApprenticeshipEarnings.TestHelpers;
-
-namespace SFA.DAS.Funding.ApprenticeshipEarnings.AcceptanceTests
+﻿namespace SFA.DAS.Funding.ApprenticeshipEarnings.AcceptanceTests
 {
     public class TestContext : IDisposable
     {
-        public TestFunction TestFunction { get; set; }
-        public SqlDatabase SqlDatabase { get; set; }
+        public TestFunction? TestFunction { get; set; }
+
         public void Dispose()
         {
-            //throw new NotImplementedException(); //todo stop function?
+            TestFunction?.Dispose();
         }
     }
 }
