+++ src/AcceptanceTests/TestFunction.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
+﻿using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -12,39 +11,25 @@ namespace SFA.DAS.Funding.ApprenticeshipEarnings.AcceptanceTests;
 
 public class Settings
 {
-    public string EnvironmentName { get; set; }
     public string AzureWebJobsStorage { get; set; }
     public string NServiceBusConnectionString { get; set; }
     public string TopicPath { get; set; }
     public string QueueName { get; set; }
-    public string LearningTransportStorageDirectory { get; set; }
 }
 
 public class TestFunction : IDisposable
 {
-    private static readonly string NServiceBusConnectionString = Environment.GetEnvironmentVariable("NServiceBusConnectionString", EnvironmentVariableTarget.Process);
-    private static readonly string AzureWebJobsStorage = Environment.GetEnvironmentVariable("AzureWebJobsStorage", EnvironmentVariableTarget.Process);
-    private static readonly string TopicPath = Environment.GetEnvironmentVariable("TopicPath", EnvironmentVariableTarget.Process);
-    private static readonly string QueueName = Environment.GetEnvironmentVariable("QueueName", EnvironmentVariableTarget.Process);
-
-    public const int BusinessCentralPaymentRequestsLimit = 1;
     private readonly TestContext _testContext;
-    private readonly Dictionary<string, string> _appConfig;
     private readonly IHost _host;
-    private readonly OrchestrationData _orchestrationData;
-    private readonly Settings _settings;
-    private bool isDisposed;
+    private bool _isDisposed;
 
     private IJobHost Jobs => _host.Services.GetService<IJobHost>();
     public string HubName { get; }
-    public HttpResponseMessage LastResponse => ResponseObject as HttpResponseMessage;
-    public ObjectResult HttpObjectResult => ResponseObject as ObjectResult;
-    public object ResponseObject { get; private set; }
 
     public TestFunction(TestContext testContext, string hubName)
     {
         HubName = hubName;
-        _orchestrationData = new OrchestrationData();
+        var orchestrationData = new OrchestrationData();
 
         _testContext = testContext;
 
@@ -55,21 +40,20 @@ public class TestFunction : IDisposable
             .AddEnvironmentVariables()
             .Build();
 
-        _settings = new Settings();
+        var settings = new Settings();
 
-        config.Bind(_settings);
+        config.Bind(settings);
 
-        _appConfig = new Dictionary<string, string>{
+        var appConfig = new Dictionary<string, string>{
             { "EnvironmentName", "LOCAL_ACCEPTANCE_TESTS" },
-            { "AzureWebJobsStorage", _settings.AzureWebJobsStorage },
-            { "NServiceBusConnectionString", _settings.NServiceBusConnectionString ?? "UseLearningEndpoint=true" },
-            { "TopicPath", _settings.TopicPath },
-            { "QueueName", _settings.QueueName }
+            { "AzureWebJobsStorage", settings.AzureWebJobsStorage },
+            { "NServiceBusConnectionString", settings.NServiceBusConnectionString ?? "UseLearningEndpoint=true" },
+            { "TopicPath", settings.TopicPath },
+            { "QueueName", settings.QueueName }
         };
 
         _testContext = testContext;
 
-        //Microsoft.Extensions.Hosting.WebJobsHostBuilderExtensions
         Environment.SetEnvironmentVariable("AzureWebJobsStorage", "UseDevelopmentStorage=true", EnvironmentVariableTarget.Process);
         Environment.SetEnvironmentVariable("NServiceBusConnectionString", "UseLearningEndpoint=true", EnvironmentVariableTarget.Process);
         Environment.SetEnvironmentVariable("ApplicationSettings:NServiceBusConnectionString", "UseLearningEndpoint=true", EnvironmentVariableTarget.Process);
@@ -80,13 +64,9 @@ public class TestFunction : IDisposable
             .ConfigureAppConfiguration(a =>
             {
                 a.Sources.Clear();
-                a.AddInMemoryCollection(_appConfig);
+                a.AddInMemoryCollection(appConfig);
             })
             .ConfigureWebJobs(builder => builder
-                //.AddHttp(options => options.SetResponse = (request, o) =>
-                //{
-                //    ResponseObject = o;
-                //}) todo what is this used for and do we need it, seemingly no longer supported in latest version of Microsoft.Extensions.Hosting
                 .AddDurableTask(options =>
                 {
                     options.HubName = HubName;
@@ -95,47 +75,21 @@ public class TestFunction : IDisposable
                     options.ExtendedSessionsEnabled = false;
                     options.StorageProvider["maxQueuePollingInterval"] = new TimeSpan(0, 0, 0, 0, 500);
                     options.StorageProvider["partitionCount"] = 1;
-                    //options.NotificationUrl = new Uri("localhost:7071"); todo again no longer supported
-#pragma warning disable S125 // Sections of code should not be commented out
-                    //options.StorageProvider["controlQueueBatchSize"] = 5;
-                    //options.HttpSettings.DefaultAsyncRequestSleepTimeMilliseconds = 500;
-                    //options.MaxConcurrentActivityFunctions = 10;
-                    //options.MaxConcurrentOrchestratorFunctions = 5;
-#pragma warning restore S125
                 })
                 .AddAzureStorageCoreServices()
                 .ConfigureServices(s =>
                 {
                     s.Configure<ApplicationSettings>(a =>
                     {
-                        a.AzureWebJobsStorage = _appConfig["AzureWebJobsStorage"];
-                        a.QueueName = _appConfig["QueueName"];
-                        a.TopicPath = _appConfig["TopicPath"];
-                        a.ServiceBusConnectionString = _appConfig["NServiceBusConnectionString"];
+                        a.AzureWebJobsStorage = appConfig["AzureWebJobsStorage"];
+                        a.QueueName = appConfig["QueueName"];
+                        a.TopicPath = appConfig["TopicPath"];
+                        a.ServiceBusConnectionString = appConfig["NServiceBusConnectionString"];
                     });
-
-                    
 
                     new Startup().Configure(builder);
 
-
-                    //todo what if any extra configure calls do we need here
-                    //s.Configure<MatchedLearnerApi>(l =>
-                    //{
-                    //    l.ApiBaseUrl = _testContext.LearnerMatchApi.BaseAddress;
-                    //    l.Identifier = "";
-                    //    l.Version = "1.0";
-                    //});
-
-                    //s.Configure<BusinessCentralApiClient>(c =>
-                    //{
-                    //    c.ApiBaseUrl = _testContext.PaymentsApi.BaseAddress;
-                    //    c.PaymentRequestsLimit = BusinessCentralPaymentRequestsLimit;
-                    //});
-
-                    //s.AddSingleton<IDistributedLockProvider, TestDistributedLockProvider>();
-                    s.AddSingleton(typeof(IOrchestrationData), _orchestrationData);
-                    //s.Decorate(typeof(ICommandHandler<>), typeof(CommandHandlerWithTimings<>));
+                    s.AddSingleton(typeof(IOrchestrationData), orchestrationData);
                 })
             )
             .ConfigureServices(s =>
@@ -157,39 +111,7 @@ public class TestFunction : IDisposable
             throw new Exception($"Failed to start test function host within {timeout.Seconds} seconds.  Check the AzureStorageEmulator is running. ");
         }
     }
-
-    public Task Start(OrchestrationStarterInfo starter, bool throwIfFailed = true)
-    {
-        return Jobs.Start(starter, throwIfFailed);
-    }
-
-    public async Task<ObjectResult> CallEndpoint(EndpointInfo endpoint)
-    {
-        await Jobs.Start(endpoint);
-        return ResponseObject as ObjectResult;
-    }
-
-
-    //todo I think this object might be different for us to just returning the raw string for now
-    public async Task<string> GetOrchestratorStartResponse()
-    {
-        var responseString = await LastResponse.Content.ReadAsStringAsync();
-        //var responseValue = JsonConvert.DeserializeObject<OrchestratorStartResponse>(responseString);
-        return responseString;
-    }
-    //public async Task<OrchestratorStartResponse> GetOrchestratorStartResponse()
-    //{
-    //    var responseString = await LastResponse.Content.ReadAsStringAsync();
-    //    var responseValue = JsonConvert.DeserializeObject<OrchestratorStartResponse>(responseString);
-    //    return responseValue;
-    //}
-
-    public async Task<DurableOrchestrationStatus> GetStatus(string instanceId)
-    {
-        await Jobs.RefreshStatus(instanceId);
-        return _orchestrationData.Status;
-    }
-
+    
     public async Task DisposeAsync()
     {
         await Jobs.StopAsync();
@@ -204,13 +126,13 @@ public class TestFunction : IDisposable
 
     protected virtual void Dispose(bool disposing)
     {
-        if (isDisposed) return;
+        if (_isDisposed) return;
 
         if (disposing)
         {
             _host.Dispose();
         }
 
-        isDisposed = true;
+        _isDisposed = true;
     }
 }
