+++ src/Infrastructure/Configuration/ApplicationSettings.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SFA.DAS.Funding.ApprenticeshipEarnings.Infrastructure.Configuration
+﻿namespace SFA.DAS.Funding.ApprenticeshipEarnings.Infrastructure.Configuration
 {
     public class ApplicationSettings
     {
